+++ PWG/EMCAL/EMCALbase/AliAnalysisTaskEmcalLight.h
@@ -36,13 +36,8 @@ class AliAODTrack;
 #include "AliMCParticleContainer.h"
 #include "AliTrackContainer.h"
 #include "AliClusterContainer.h"
+#include "AliEmcalStringView.h"
 
-#if ROOT_VERSION_CODE > ROOT_VERSION(6,10,0) 
-#include "RStringView.h"
-#define EMCAL_STRINGVIEW const std::string_view
-#else 
-#define EMCAL_STRINGVIEW const std::string &
-#endif
 
 #include "AliAnalysisTaskSE.h"
 /**
